remove env requirement for mockshop
