+++ packages/hydrogen/src/analytics-manager/ShopifyAnalytics.tsx
@@ -29,6 +29,8 @@ import {
   ComponentizableCartLine,
   Maybe,
 } from '@shopify/hydrogen-react/storefront-api-types';
+import invariant from 'tiny-invariant';
+import {errorOnce} from '../utils/warning';
 
 function getCustomerPrivacyRequired() {
   const customerPrivacy = getCustomerPrivacy();
@@ -41,6 +43,11 @@ function getCustomerPrivacyRequired() {
 
   return customerPrivacy;
 }
+
+function messageOnError(field: string) {
+  return `[h2:error:Analytics.Provider] - ${field} is required`;
+}
+
 
 /**
  * This component is responsible for sending analytics events to Shopify.
@@ -55,11 +62,41 @@ export function ShopifyAnalytics({
   consent,
   onReady,
   domain,
+  disableThrowOnError,
+  isMockShop
 }: {
   consent: AnalyticsProviderProps['consent'];
   onReady: () => void;
   domain?: string;
+  disableThrowOnError: boolean;
+  isMockShop: boolean;
 }) {
+
+  // If mock shop is used, log error instead of throwing
+  if (isMockShop) {
+    errorOnce('[h2:error:Analytics.Provider] - Mock shop is used. Analytics will not work properly.')
+  } else {
+    if (!consent.checkoutDomain) {
+      const errorMsg = messageOnError('consent.checkoutDomain');
+      if (disableThrowOnError) {
+        // eslint-disable-next-line no-console
+        console.error(errorMsg);
+      } else {
+        invariant(false, errorMsg);
+      }
+    }
+
+    if (!consent.storefrontAccessToken) {
+      const errorMsg = messageOnError('consent.storefrontAccessToken');
+      if (disableThrowOnError) {
+        // eslint-disable-next-line no-console
+        console.error(errorMsg);
+      } else {
+        invariant(false, errorMsg);
+      }
+    }
+  }
+
   const {subscribe, register, canTrack} = useAnalytics();
   const [shopifyReady, setShopifyReady] = useState(false);
   const [privacyReady, setPrivacyReady] = useState(false);
@@ -78,15 +115,14 @@ export function ShopifyAnalytics({
   };
 
   useCustomerPrivacy({
-    ...consent,
+    checkoutDomain: isMockShop ? 'mock.shop' : consent.checkoutDomain,
+    storefrontAccessToken: isMockShop ? 'abcdefghijklmnopqrstuvwxyz123456' :consent.storefrontAccessToken,
     onVisitorConsentCollected: setCustomerPrivacyReady,
     onReady: () => {
       // Set customer privacy ready 3 seconds after load
       setTimeout(setCustomerPrivacyReady, 3000);
     },
   });
-
-  console.log('Shopify analytics:', domain)
 
   useShopifyCookies({
     hasUserConsent: canTrack(),
