+++ templates/skeleton/.env
@@ -1,8 +1,3 @@
 # These variables are only available locally in MiniOxygen
 
 SESSION_SECRET="foobar"
-
-# Mock.shop ENV
-PUBLIC_STORE_DOMAIN="mock.shop"
-PUBLIC_CHECKOUT_DOMAIN="mock.shop"
-PUBLIC_STOREFRONT_API_TOKEN="abcdefghijklmnopqrstuvwxyz123456"
